SCRUM-27: Clean up files (delete commented lines etc.) and outsourced graph interfaces in separate file
