+++ src/app/services/calculate-coordinates.service.spec.ts
@@ -1,0 +1,11 @@
+import { TestBed } from '@angular/core/testing';
+
+import { CalculateCoordinatesService } from './calculate-coordinates.service';
+
+describe('CalculateCoordinatesService', () => {
+    let service: CalculateCoordinatesService;
+
+    beforeEach(() => {});
+
+    it('should be created', () => {});
+});
