+++ src/app/services/calculate-coordinates.service.ts
@@ -3,28 +3,7 @@ import { Dfg } from '../classes/dfg/dfg';
 import { Activities, Activity } from '../classes/dfg/activities';
 import { DrawingAreaComponent } from '../components/drawing-area';
 import { BehaviorSubject, last, Observable } from 'rxjs';
-
-export interface Graph {
-    nodes: Node[];
-    edges: Edge[];
-}
-
-export interface Node {
-    id: string;
-    x: number;
-    y: number;
-}
-
-interface Edge {
-    source: Node;
-    target: Node;
-}
-
-interface StackElement {
-    activity: Activity;
-    source_x: number;
-    source_y: number;
-}
+import { Graph, Node, Edge, StackElement } from '../classes/graph';
 
 @Injectable({
     providedIn: 'root',
@@ -45,23 +24,26 @@ export class CalculateCoordinatesService {
     }
 
     public calculateCoordinates(dfg: Dfg): void {
-        //const map: Map<String, Array<Number>> = new Map();
-        //const graphStack = this.createGraphStack(dfg);
         const nodes: Array<Node> = [];
+        const edges: Array<Edge> = [];
+
         let yCoordinate: number = 100;
+        let lastX: number = 100;
 
         const playActivity: Activity = dfg.activities.getActivityByName('play');
         nodes.push({ id: playActivity.name, x: 100, y: 100 });
         const neighbours: Activities =
             dfg.arcs.calculateNextActivities(playActivity);
 
+        /*
+        Die Koordinaten des DFG werden mittels Tiefensuche von Graphen (DFS) die Koordinaten berechnet.
+        Der Stack enthält dabei die Elemente des Graphen, die schon gefunden wurden, deren Koordinanten aber noch berechnet werden müssen.
+        */
         const stack: Array<StackElement> = neighbours.activities.map(
             (neighbour) => {
                 return { activity: neighbour, source_x: 100, source_y: 100 };
             },
         );
-
-        let lastX: number = 100;
 
         while (stack.length > 0) {
             const stackElement: StackElement = stack.pop()!;
@@ -100,8 +82,6 @@ export class CalculateCoordinatesService {
             }
         }
 
-        const edges: Array<Edge> = [];
-
         dfg.arcs.arcs.forEach((arc) => {
             const startNode: Node = nodes.find(
                 (node) => arc.getStart().name === node.id,
@@ -114,74 +94,5 @@ export class CalculateCoordinatesService {
         });
 
         this._graph$.next({ nodes: nodes, edges: edges });
-        console.log(dfg.arcs);
-        console.log(edges);
-        console.log(nodes);
-
-        // console.log(graphStack);
-
-        // let lastX = 10;
-        // let lastY = 10;
-        // let stack = [];
-
-        // graphStack.forEach((neighbour, activity) => {
-        //     if (activity == 'play') {
-        //         map.set(activity, [lastX, lastY]);
-        //         let neighbourSize = neighbour.length;
-        //         lastX = lastX + 10;
-        //         //queue.push(neighbour);
-
-        //         neighbour.forEach((neighbour1, index) => {
-        //             map.set(neighbour1.name, [
-        //                 lastX,
-        //                 lastY + lastY * (index / neighbourSize),
-        //             ]);
-        //         });
-        //     }
-
-        //     // while (queue.length != 0) {
-
-        //     // }
-        //     // } else if (activity != 'stop') {
-        //     //     console.log(activity);
-        //     //     console.log('letzes X: ' + lastX);
-        //     //     if (!map.has(activity)) {
-        //     //         map.set(activity, [lastX, lastY]);
-        //     //         let neighbourSize = neighbour.length;
-        //     //         lastX = lastX + 10;
-
-        //     //         neighbour.forEach((neighbour1, index) => {
-        //     //             map.set(neighbour1.name, [
-        //     //                 lastX,
-        //     //                 lastY + lastY * (index / neighbourSize),
-        //     //             ]);
-        //     //         });
-        //     //     }
-        //     //     lastX = lastX + 10;
-        //     // }
-
-        //     lastX = lastX + 10;
-
-        // lastX = newX
-        // });
-
-        // return map;
-    }
-
-    public createGraphStack(dfg: Dfg): Map<String, Array<Activity>> {
-        const graphStack = new Map<String, Array<Activity>>();
-
-        dfg.activities.activities.forEach((act) => {
-            graphStack.set(act.name, []);
-
-            const neighbours: Activities =
-                dfg.arcs.calculateNextActivities(act);
-
-            neighbours.activities.forEach((act2) => {
-                graphStack.get(act.name)?.push(act2);
-            });
-        });
-
-        return graphStack;
     }
 }
